+++ NEWS.md
@@ -2,6 +2,9 @@
 
 #### ✨ features and improvements
 
+  * Attempts to fix broken geometries with `sf::st_make_valid()` (both with and 
+    without {s2} 📦 support), thus eliminating {maptools} 📦 dependency (#139) 
+
 #### 🐛 bug fixes
 
 #### 💬 documentation etc